Better error handling: stop container on test fail
Use the bash equivalent of try/finally to, i.e. trap <tidy> EXIT, to
detect whether we had successfully started a container (by looking
in the variable) and if so, stop it.

This way, the test (psql command) is still the last call, so the
script as a whole still returns a failure when the test fail.