Implement a healthcheck so we can establish db is running
It seems the base image starts the db server, then apply our
custom configuration, then restarts the db server. The restart
happens asynchronously, i.e. the 'docker run' finishes before and
the containers appear 'running' then the db server restarts.

This can cause confusion for the consumers as they *are* able to
connect then the connection is reset.

It seems the best we can do (unless we want to tinker with the
entry point of the base image?) is to make use of docker's
HEALTHCHECK, which doesn't prevent the above behaviour but at
least gives us a way to prod the container to wait until the
restart has already happened.